backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0.0,<2.20.0
langgraph>=1.0.0
langchain-core>=1.0.0
httpx>=0.25.0
pydantic>=2.8.0
websockets>=15.0.0

